pyproject.toml
[tool.poetry]
name = "prefect-pipeline"
version = "0.1.0"
description = ""
authors = ["Ivan Reshetnikov <resivalex@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
prefect = "^2.10.17"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
